moved get_emotions_from_dict method and refactored concept code and boosted concepts in query
