+++ flask_app.py
@@ -5,7 +5,7 @@ from flask import (
     redirect, request, url_for
 )
 from TextPersonality import NLUPersonalityInterface
-from StoreChars import DiscoveryCharDatabase
+from StoreChars import DiscoveryCharDatabase, get_emotions_from_dict
 from werkzeug.exceptions import InternalServerError
 from chart import create_figure
 import time
@@ -71,15 +71,6 @@ def handle_500(e):
     # wrapped unhandled error
     return render_template("500_unhandled.html", e=original), 500
 
-
-def get_emotions_from_dict(personality):
-    emotions_dict = dict()
-    emotions_dict["sadness"] = personality.pop("sadness", 0)
-    emotions_dict["joy"] = personality.pop("joy", 0)
-    emotions_dict["fear"] = personality.pop("fear", 0)
-    emotions_dict["disgust"] = personality.pop("disgust", 0)
-    emotions_dict["anger"] = personality.pop("anger", 0)
-    return emotions_dict, personality
 
 
 app.run()
